validar

+++ vistas/js/ordenes.js
@@ -145,25 +145,7 @@ function guardar_detalles(datos) {
 		deta.idOrden=datos;
 		ordenSer=deta.idServicio+""+deta.idOrden;
 		ordenSer=ordenSer.replace(" ", "");
-		/*$.ajax({
-			url: "../ajax/Orden.php?op=detallesDetalles",
-			type: "POST",
-			data: deta,
-			contentType: false,
-			processData: false,
-
-			success: function(datos){
-
-				console.log(datos); 
-
-				
-				servicioEmpleada(datos,deta.idServicio);
-				
-				
-				
-				
-			}
-		});*/
+		
 		$.post("../ajax/Orden.php?op=detallesDetalles",{"arregloDetalle":deta},function(datas){
 
 		});
@@ -184,7 +166,7 @@ function servicioEmpleada(data,servicio) {
 				empleada.ordenServicio=String(data);
 				$.post("../ajax/Orden.php?op=detallesEmpleada",{"arregloEmpleada":empleada},function(data){
 		
-					alert(data);
+					bootbox.alert(data);
 				});
 			} 			
 		});
@@ -238,7 +220,7 @@ function dibujartabla(arregloDetalle){
 		tdEliminar.appendChild(botonEliminar);
 		fila.appendChild(tdEliminar);
 		botonEliminar.onclick=()=>{
-			eliminarIten(deta.idServicio);
+			eliminarIten(deta.idServicio,deta.nombreServi);
 		};
 		
 		cuerpotabla.appendChild(fila);
@@ -248,8 +230,8 @@ function dibujartabla(arregloDetalle){
 function dibujartablaE(arregloEmpleada){
 	cuerpotablaE.innerHTML="";
 	arregloEmpleada.forEach((data)=>{
-		console.log("srevi",data.idServicio);
-		console.log("imput",$("#idser").val());
+		console.log("1234",data.idServicio);
+		console.log("123",$("#idser").val());
 		if (data.idServicio===$("#idser").val()) {
 			let fila = document.createElement("tr");
 			fila.innerHTML=`
@@ -275,27 +257,61 @@ function dibujartablaE(arregloEmpleada){
 }
 
 function agregar_detalles(){
-		const objDetalles={
-			precio: precio.value,
-			idServicio:idServicio.value,
-			descripcion:descripcion.value,
-			nombreServi:nombreServi.value,
+	if ($("#idServicio").val()!==null) {
+		console.log("descripcion", $("#descripcion").val());
+		if ($("#descripcion").val()!='') {
+			const objDetalles={
+				precio: precio.value,
+				idServicio:idServicio.value,
+				descripcion:descripcion.value,
+				nombreServi:nombreServi.value,
+			}
+			$("#idServicio").val('default').selectpicker("refresh");
+			LlenarDetalles(objDetalles);
+			dibujartabla(arregloDetalle);
+			console.log(arregloDetalle);
+			document.getElementById("descripcion").style.border = 0;
+		} else {
+			document.getElementById("descripcion").style.borderColor = "red";
+			swal.fire(
+				"¡Ingrese Descripcion!",
+				"No se puede agreagar un servicio sin descripcion",
+				"error"
+				);
+		
 		}
-		LlenarDetalles(objDetalles);
-		dibujartabla(arregloDetalle);
-		console.log(arregloDetalle);
+		
+	} else {
+		swal.fire(
+			"¡Seleccione un servicio!",
+			"No se puede agreagar un servicio sin datos",
+			"error"
+			);
+	
+	}
 }//fin guardar y editar
 
 function agregar_empleadas(){
 
-	const objEmpleada={
-		idServicio:idServicio.value,
-		empleada:idEmpleada.value,
-		nombreE:nombreE,	
+	if ($("#idEmpleada").val()!==null) {
+		const objEmpleada={
+			idServicio:idServicio.value,
+			empleada:idEmpleada.value,
+			nombreE:nombreE,	
+		}
+		//$("#idEmpleada").val('default').selectpicker("refresh");
+		LlenarEmpleada(objEmpleada);
+
+		dibujartablaE(arregloEmpleada);
+		console.log("0000000000",arregloEmpleada);
+	} else {
+		swal.fire(
+			"¡Seleccione un empleado!",
+			"No se puede agreagar un empleado sin datos",
+			"error"
+		);
 	}
-	LlenarEmpleada(objEmpleada);
-	dibujartablaE(arregloEmpleada);
-	console.log(arregloEmpleada);
+	
 }
 function abrirModal(idServicio,nombreServi) {
 	
@@ -318,7 +334,13 @@ function LlenarDetalles(objDetalles){
 
 	if (!result) {
 		arregloDetalle.push(objDetalles);
-	} 
+	} else {
+		swal.fire(
+			"¡Registro duplicado!",
+			"El servicio ya esta registrado para esta orden",
+			"error"
+		  );
+	}
 	
 	
 }
@@ -339,21 +361,51 @@ function LlenarEmpleada(objEmpleada){
 	if (!result) {
 		arregloEmpleada.push(objEmpleada);
 	} else{
-		alert("Emplado ya registrado para este servicio");
+		//bootbox.alert("Emplado ya registrado para este servicio");
+		swal.fire(
+			"¡Registro duplicado!",
+			"Emplado ya registrado para este servicio",
+			"error"
+		  );
 	}
 	
 	
 }
-function eliminarIten(id){
-	arregloDetalle= arregloDetalle.filter((detalle)=>{
-		if (+id !== +detalle.idServicio) {
-			return detalle;
-			
-		}else{
-			eliminarEmpleadaS(id);
-		}
+function eliminarIten(id,nombre){
+	swal.fire({
+		title: "¿Esta seguro(a) de eliminar este servico: "+nombre+", de esta orden?",
+		type: "question",
+		showCancelButton: true,
+		confirmButtonText: "Eliminar",
+		confirmButtonColor: "#3085d6",
+		cancelButtonColor: "#ca5939"
+	})
+	.then(result => {
+		
+		 if (result.value) {
+			arregloDetalle= arregloDetalle.filter((detalle)=>{
+				if (+id !== +detalle.idServicio) {
+					return detalle;
+					
+				}else{
+					
+					eliminarEmpleadaS(id);
+				}
+			});
+			dibujartabla(arregloDetalle);
+
+		swal.fire(
+		  "Se eliminino el servicio: "+nombre,
+
+		  "¡Aviso!",
+		  "success"
+		);
+	  }
 	});
-	dibujartabla(arregloDetalle);
+
+
+	//
+	
 }
 function eliminarItenE(id){
 	arregloEmpleada= arregloEmpleada.filter((detalle)=>{
